change structure, add compressed option to ingest
