+++ src/main/java/org/broadinstitute/hellbender/tools/variantdb/IngestConstants.java
@@ -1,4 +1,4 @@
-package org.broadinstitute.hellbender.tools.variantdb.ingest;
+package org.broadinstitute.hellbender.tools.variantdb;
 
 public class IngestConstants {
