+++ src/main/java/org/broadinstitute/hellbender/tools/variantdb/arrays/ArrayMetadataTsvCreator.java
@@ -1,22 +1,21 @@
-package org.broadinstitute.hellbender.tools.variantdb.ingest;
+package org.broadinstitute.hellbender.tools.variantdb.arrays;
 
 import org.apache.commons.lang3.StringUtils;
 import org.broadinstitute.hellbender.exceptions.UserException;
-import org.broadinstitute.hellbender.tools.variantdb.ingest.nextgen.PetTsvCreator;
+import org.broadinstitute.hellbender.tools.variantdb.IngestConstants;
+import org.broadinstitute.hellbender.tools.variantdb.nextgen.PetTsvCreator;
 import org.broadinstitute.hellbender.utils.SimpleInterval;
 import org.broadinstitute.hellbender.utils.Utils;
 import org.broadinstitute.hellbender.utils.tsv.SimpleXSVWriter;
 
-import java.io.File;
 import java.io.IOException;
-import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class MetadataTsvCreator {
+public class ArrayMetadataTsvCreator {
 
     private SimpleXSVWriter sampleMetadataWriter = null;
 
@@ -25,70 +24,40 @@ public class MetadataTsvCreator {
      */
     public enum HeaderFieldEnum {
         sample_name,
-        sample_id,
-        interval_list_blob,
-        inferred_state,
-    }
-
-    private List<String> createSampleListRow(
-            String sampleName,
-            String sampleId,
-            String intervalListBlob,
-            PetTsvCreator.GQStateEnum inferredMissingState
-    ) {
-
-        List<String> row = new ArrayList<>();
-        row.add(sampleName);
-        row.add(sampleId);
-        row.add(intervalListBlob);
-        if (inferredMissingState == null) {
-            row.add("");
-        } else {
-            row.add(inferredMissingState.getValue());
-        }
-
-        return row;
+        sample_id
     }
 
     public static List<String> getHeaders() {
-        return Arrays.stream(MetadataTsvCreator.HeaderFieldEnum.values()).map(String::valueOf).collect(Collectors.toList());
+        return Arrays.stream(ArrayMetadataTsvCreator.HeaderFieldEnum.values()).map(String::valueOf).collect(Collectors.toList());
     }
 
-//    public void createRow(String sampleName, String sampleId) {
-//        createRow(sampleName, sampleId, null, PetTsvCreator.GQStateEnum.SIXTY);
-//
-//    }
-//
-    public void createRow(String sampleName, String sampleId, String tableNumberPrefix, List<SimpleInterval> userIntervals, PetTsvCreator.GQStateEnum gqStateToIgnore) {
+    public void createRow(String sampleName, String sampleId, String tableNumberPrefix) {
         // if the metadata tsvs don't exist yet -- create them
         try {
             // Create a metadata file to go into the metadata dir for _this_ sample
             // TODO--this should just be one file per sample set?
             final String sampleMetadataName = IngestConstants.metadataFilePrefix + tableNumberPrefix + sampleName + IngestConstants.FILETYPE;
             // write header to it
-            List<String> sampleListHeader = MetadataTsvCreator.getHeaders();
+            List<String> sampleListHeader = ArrayMetadataTsvCreator.getHeaders();
             sampleMetadataWriter = new SimpleXSVWriter(Paths.get(sampleMetadataName), IngestConstants.SEPARATOR);
             sampleMetadataWriter.setHeaderLine(sampleListHeader);
-            String intervalListMd5 = "NA";
 
-            if (userIntervals != null) {
-                // write values
-                List<String> intervalList = userIntervals.stream().map(interval -> interval.toString())
-                        .collect(Collectors.toList());
-                String intervalListBlob = StringUtils.join(intervalList, ", ");
-                intervalListMd5 = Utils.calcMD5(intervalListBlob);
-            }
             final List<String> TSVLineToCreateSampleMetadata = createSampleListRow(
                     sampleName,
-                    sampleId,
-                    intervalListMd5,
-                    gqStateToIgnore);
+                    sampleId);
             sampleMetadataWriter.getNewLineBuilder().setRow(TSVLineToCreateSampleMetadata).write();
 
         } catch (final IOException e) {
             throw new UserException("Could not create sample metadata outputs", e);
         }
 
+    }
+
+    private List<String> createSampleListRow(String sampleName, String sampleId) {
+        List<String> row = new ArrayList<>();
+        row.add(sampleName);
+        row.add(sampleId);
+        return row;
     }
 
     public void closeTool() {
