+++ src/main/java/org/broadinstitute/hellbender/tools/variantdb/arrays/RawArrayTsvCreator.java
@@ -1,4 +1,4 @@
-package org.broadinstitute.hellbender.tools.variantdb.ingest.arrays;
+package org.broadinstitute.hellbender.tools.variantdb.arrays;
 import htsjdk.variant.variantcontext.VariantContext;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -6,7 +6,7 @@ import org.broadinstitute.hellbender.engine.FeatureContext;
 import org.broadinstitute.hellbender.engine.ReadsContext;
 import org.broadinstitute.hellbender.engine.ReferenceContext;
 import org.broadinstitute.hellbender.exceptions.UserException;
-import org.broadinstitute.hellbender.tools.variantdb.ingest.IngestConstants;
+import org.broadinstitute.hellbender.tools.variantdb.IngestConstants;
 import org.broadinstitute.hellbender.utils.tsv.SimpleXSVWriter;
 
 
@@ -24,6 +24,7 @@ public final class RawArrayTsvCreator {
     private SimpleXSVWriter rawArrayWriter = null;
     private final String sampleId;
     private final Map<String, ProbeInfo> probeDataByName;
+    private final boolean useCompressedData;
     private static String RAW_FILETYPE_PREFIX = "raw_";
 
     enum GT_encoding {
@@ -41,9 +42,10 @@ public final class RawArrayTsvCreator {
         }
     }
 
-    public RawArrayTsvCreator(final String sampleName, final String sampleId, final String tableNumberPrefix, final Map<String, ProbeInfo> probeDataByName) {
+    public RawArrayTsvCreator(final String sampleName, final String sampleId, final String tableNumberPrefix, final Map<String, ProbeInfo> probeDataByName, boolean useCompressedData) {
         this.sampleId = sampleId;
         this.probeDataByName = probeDataByName;
+        this.useCompressedData = useCompressedData;
         try {
             // Create a raw file to go into the raw dir for _this_ sample
             final String rawOutputName = RAW_FILETYPE_PREFIX + tableNumberPrefix + sampleName  + IngestConstants.FILETYPE;
@@ -58,7 +60,6 @@ public final class RawArrayTsvCreator {
 
     public List<String> createRow(final VariantContext variant, final String sampleId) {
         List<String> row = new ArrayList<>();
-        row.add(sampleId);
         String rsid = variant.getID();
         if (rsid == null) {
             throw new IllegalStateException("Cannot be missing required value for site_name"); // TODO, should this be UserException too?
@@ -67,10 +68,12 @@ public final class RawArrayTsvCreator {
         if (probeInfo == null) {
             logger.warn("no probe found for variant with ID: " + variant.getID() + "\t" + variant);
         } else {
-            for (final RawArrayFieldEnum fieldEnum : RawArrayFieldEnum.values()) {
-                if (!fieldEnum.equals(RawArrayFieldEnum.sample_id)) {
-                    row.add(fieldEnum.getColumnValue(variant, probeInfo, sampleId));
-                }
+            RawArrayFieldEnum[] fields = RawArrayFieldEnum.getCompressedRawArrayFieldEnums();
+            if (!useCompressedData) {
+                fields = RawArrayFieldEnum.getUncompressedRawArrayFieldEnums();
+            }
+            for (final RawArrayFieldEnum fieldEnum : fields) {
+                row.add(fieldEnum.getColumnValue(variant, probeInfo, sampleId));
             }
         }
         return row;
@@ -78,11 +81,6 @@ public final class RawArrayTsvCreator {
 
     public static List<String> getHeaders() {
         return Arrays.stream(RawArrayFieldEnum.values()).map(String::valueOf).collect(Collectors.toList());
-    }
-
-    public List<List<String>> createRows(long start, long end, VariantContext variant, String sampleId) {
-        // this doesn't apply to arrays
-        throw new UserException.UnimplementedFeature("Not implemented.");
     }
 
     public void apply(final VariantContext variant, final ReadsContext readsContext, final ReferenceContext referenceContext, final FeatureContext featureContext) {
