--- src/main/java/org/broadinstitute/hellbender/tools/variantdb/ingestutils/pet_schema.json
@@ -1,20 +1,0 @@
-[
- {
-   "description": "[DESCRIPTION]",
-   "name": "position",
-   "type": "Integer",
-   "mode": "Required"
- },
- {
-   "description": "[DESCRIPTION]",
-   "name": "sample",
-   "type": "String",
-   "mode": "Required"
- },
-  {
-   "description": "[DESCRIPTION]",
-   "name": "state",
-   "type": "String",
-   "mode": "Required"
- }
-]
