--- src/main/java/org/broadinstitute/hellbender/tools/variantdb/ingestutils/sample_list_schema.json
@@ -1,20 +1,0 @@
-[
-  {
-    "description": "[DESCRIPTION]",
-    "name": "sample",
-    "type": "String",
-    "mode": "Required"
-  },
-  {
-    "description": "[DESCRIPTION]",
-    "name": "interval_list_blob",
-    "type": "String",
-    "mode": "Required"
-  },
-  {
-    "description": "[DESCRIPTION]",
-    "name": "inferred_state",
-    "type": "String",
-    "mode": "Required"
-  }
-]
