+++ src/main/java/org/broadinstitute/hellbender/tools/variantdb/nextgen/MetadataTsvCreator.java
@@ -1,9 +1,8 @@
-package org.broadinstitute.hellbender.tools.variantdb.ingest.arrays;
+package org.broadinstitute.hellbender.tools.variantdb.nextgen;
 
 import org.apache.commons.lang3.StringUtils;
 import org.broadinstitute.hellbender.exceptions.UserException;
-import org.broadinstitute.hellbender.tools.variantdb.ingest.IngestConstants;
-import org.broadinstitute.hellbender.tools.variantdb.ingest.nextgen.PetTsvCreator;
+import org.broadinstitute.hellbender.tools.variantdb.IngestConstants;
 import org.broadinstitute.hellbender.utils.SimpleInterval;
 import org.broadinstitute.hellbender.utils.Utils;
 import org.broadinstitute.hellbender.utils.tsv.SimpleXSVWriter;
@@ -15,7 +14,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class ArrayMetadataTsvCreator {
+public class MetadataTsvCreator {
 
     private SimpleXSVWriter sampleMetadataWriter = null;
 
@@ -24,40 +23,70 @@ public class ArrayMetadataTsvCreator {
      */
     public enum HeaderFieldEnum {
         sample_name,
-        sample_id
+        sample_id,
+        interval_list_blob,
+        inferred_state,
+    }
+
+    private List<String> createSampleListRow(
+            String sampleName,
+            String sampleId,
+            String intervalListBlob,
+            PetTsvCreator.GQStateEnum inferredMissingState
+    ) {
+
+        List<String> row = new ArrayList<>();
+        row.add(sampleName);
+        row.add(sampleId);
+        row.add(intervalListBlob);
+        if (inferredMissingState == null) {
+            row.add("");
+        } else {
+            row.add(inferredMissingState.getValue());
+        }
+
+        return row;
     }
 
     public static List<String> getHeaders() {
-        return Arrays.stream(ArrayMetadataTsvCreator.HeaderFieldEnum.values()).map(String::valueOf).collect(Collectors.toList());
+        return Arrays.stream(MetadataTsvCreator.HeaderFieldEnum.values()).map(String::valueOf).collect(Collectors.toList());
     }
 
-    public void createRow(String sampleName, String sampleId, String tableNumberPrefix) {
+//    public void createRow(String sampleName, String sampleId) {
+//        createRow(sampleName, sampleId, null, PetTsvCreator.GQStateEnum.SIXTY);
+//
+//    }
+//
+    public void createRow(String sampleName, String sampleId, String tableNumberPrefix, List<SimpleInterval> userIntervals, PetTsvCreator.GQStateEnum gqStateToIgnore) {
         // if the metadata tsvs don't exist yet -- create them
         try {
             // Create a metadata file to go into the metadata dir for _this_ sample
             // TODO--this should just be one file per sample set?
             final String sampleMetadataName = IngestConstants.metadataFilePrefix + tableNumberPrefix + sampleName + IngestConstants.FILETYPE;
             // write header to it
-            List<String> sampleListHeader = ArrayMetadataTsvCreator.getHeaders();
+            List<String> sampleListHeader = MetadataTsvCreator.getHeaders();
             sampleMetadataWriter = new SimpleXSVWriter(Paths.get(sampleMetadataName), IngestConstants.SEPARATOR);
             sampleMetadataWriter.setHeaderLine(sampleListHeader);
+            String intervalListMd5 = "NA";
 
+            if (userIntervals != null) {
+                // write values
+                List<String> intervalList = userIntervals.stream().map(interval -> interval.toString())
+                        .collect(Collectors.toList());
+                String intervalListBlob = StringUtils.join(intervalList, ", ");
+                intervalListMd5 = Utils.calcMD5(intervalListBlob);
+            }
             final List<String> TSVLineToCreateSampleMetadata = createSampleListRow(
                     sampleName,
-                    sampleId);
+                    sampleId,
+                    intervalListMd5,
+                    gqStateToIgnore);
             sampleMetadataWriter.getNewLineBuilder().setRow(TSVLineToCreateSampleMetadata).write();
 
         } catch (final IOException e) {
             throw new UserException("Could not create sample metadata outputs", e);
         }
 
-    }
-
-    private List<String> createSampleListRow(String sampleName, String sampleId) {
-        List<String> row = new ArrayList<>();
-        row.add(sampleName);
-        row.add(sampleId);
-        return row;
     }
 
     public void closeTool() {
